feature: combine query and lucene
